requirements.txt
django
mysqlclient
Pillow
djangorestframework
requests
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
django-jazzmin
django-js-asset
Jinja2
redis
template
bcrypt
PyMuPDF
reportlab
pdf2docx
django-environ
requests 

# pip install -r requirements.txt
# python manage.py migrate
# python manage.py runserver